Cleanup; Udpate readme for branch
