+++ reasoner/Reasoner.java
@@ -23,7 +23,6 @@ import com.vaticle.typedb.core.concurrent.producer.Producer;
 import com.vaticle.typedb.core.concurrent.producer.Producers;
 import com.vaticle.typedb.core.logic.LogicManager;
 import com.vaticle.typedb.core.logic.resolvable.Concludable;
-import com.vaticle.typedb.core.logic.resolvable.ResolvableConjunction;
 import com.vaticle.typedb.core.logic.resolvable.ResolvableDisjunction;
 import com.vaticle.typedb.core.pattern.Conjunction;
 import com.vaticle.typedb.core.pattern.Disjunction;
@@ -31,10 +30,9 @@ import com.vaticle.typedb.core.pattern.Negation;
 import com.vaticle.typedb.core.pattern.variable.Variable;
 import com.vaticle.typedb.core.reasoner.answer.Explanation;
 import com.vaticle.typedb.core.reasoner.common.ReasonerPerfCounters;
-import com.vaticle.typedb.core.reasoner.controller.ControllerRegistry;
 import com.vaticle.typedb.core.reasoner.planner.ReasonerPlanner;
 import com.vaticle.typedb.core.reasoner.v4.nodes.NodeRegistry;
-import com.vaticle.typedb.core.reasoner.v4.ReasonerProducerV4;
+import com.vaticle.typedb.core.reasoner.v4.ReasonerProducer;
 import com.vaticle.typedb.core.traversal.TraversalEngine;
 import com.vaticle.typedb.core.traversal.common.Identifier;
 import com.vaticle.typedb.core.traversal.common.Modifiers.Filter;
@@ -54,7 +52,6 @@ import static com.vaticle.typedb.common.collection.Collections.set;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Pattern.INFERENCE_INCOHERENT_MATCH_PATTERN;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Pattern.INFERENCE_INCOHERENT_MATCH_SUB_PATTERN;
-import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.UNIMPLEMENTED;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.ThingRead.SORT_ATTRIBUTE_NOT_COMPARABLE;
 import static com.vaticle.typedb.core.common.iterator.Iterators.cartesian;
 import static com.vaticle.typedb.core.common.iterator.Iterators.empty;
@@ -72,7 +69,6 @@ public class Reasoner {
     private final TraversalEngine traversalEng;
     private final ConceptManager conceptMgr;
     private final LogicManager logicMgr;
-    private final ControllerRegistry controllerRegistry;
     private final ExplainablesManager explainablesManager;
     private final ReasonerPlanner planner;
     private final ReasonerPerfCounters perfCounters;
@@ -85,14 +81,9 @@ public class Reasoner {
         this.logicMgr = logicMgr;
         this.perfCounters = new ReasonerPerfCounters(context.options().infer() && context.options().reasonerPerfCounters());
         this.planner = ReasonerPlanner.create(traversalEng, conceptMgr, logicMgr, perfCounters, context.options().explain());
-        this.controllerRegistry = new ControllerRegistry(actor(), traversalEng, conceptMgr, logicMgr, planner, perfCounters, context);
         this.explainablesManager = new ExplainablesManager();
 
         this.nodeRegistry = new NodeRegistry(actor(), perfCounters, conceptMgr, logicMgr , traversalEng, planner, context.options());
-    }
-
-    public ControllerRegistry controllerRegistry() {
-        return controllerRegistry;
     }
 
     public FunctionalIterator<? extends ConceptMap> execute(Disjunction disjunction, List<Identifier.Variable.Name> filterVars,
@@ -224,7 +215,7 @@ public class Reasoner {
     }
 
     public FunctionalIterator<ConceptMap> executeReasoner(Disjunction disjunction, Filter filter, Context.Query context) {
-        ReasonerProducerV4.Basic producer = new ReasonerProducerV4.Basic(ResolvableDisjunction.of(disjunction), filter, context.options(), nodeRegistry, explainablesManager);
+        ReasonerProducer.Basic producer = new ReasonerProducer.Basic(ResolvableDisjunction.of(disjunction), filter, context.options(), nodeRegistry, explainablesManager);
         return produce(producer, context.producer(), async1());
     }
 
@@ -320,13 +311,13 @@ public class Reasoner {
         Concludable explainableConcludable = explainablesManager.getConcludable(explainableId);
         ConceptMap explainableBounds = explainablesManager.getBounds(explainableId);
         return produce(
-                list(new ReasonerProducerV4.Explain(explainableConcludable, explainableBounds, defaultContext.options(), nodeRegistry, explainablesManager)),
+                list(new ReasonerProducer.Explain(explainableConcludable, explainableBounds, defaultContext.options(), nodeRegistry, explainablesManager)),
                 Either.first(Arguments.Query.Producer.INCREMENTAL),
                 async1()
         );
     }
 
     public void close() {
-        controllerRegistry.close();
+        nodeRegistry.close();
     }
 }
