--- reasoner/ReasonerProducer.java
@@ -1,227 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-
-package com.vaticle.typedb.core.reasoner;
-
-import com.vaticle.typedb.core.common.parameters.Options;
-import com.vaticle.typedb.core.concept.answer.ConceptMap;
-import com.vaticle.typedb.core.concurrent.actor.Actor;
-import com.vaticle.typedb.core.concurrent.producer.Producer;
-import com.vaticle.typedb.core.logic.resolvable.Concludable;
-import com.vaticle.typedb.core.logic.resolvable.ResolvableConjunction;
-import com.vaticle.typedb.core.logic.resolvable.ResolvableDisjunction;
-import com.vaticle.typedb.core.reasoner.answer.Explanation;
-import com.vaticle.typedb.core.reasoner.controller.ControllerRegistry;
-import com.vaticle.typedb.core.reasoner.processor.AbstractProcessor;
-import com.vaticle.typedb.core.traversal.common.Modifiers;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.annotation.concurrent.ThreadSafe;
-import java.util.concurrent.Executor;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import static com.vaticle.typedb.core.reasoner.ReasonerProducer.State.EXCEPTION;
-import static com.vaticle.typedb.core.reasoner.ReasonerProducer.State.FINISHED;
-import static com.vaticle.typedb.core.reasoner.ReasonerProducer.State.INIT;
-import static com.vaticle.typedb.core.reasoner.ReasonerProducer.State.INITIALISING;
-import static com.vaticle.typedb.core.reasoner.ReasonerProducer.State.PULLING;
-import static com.vaticle.typedb.core.reasoner.ReasonerProducer.State.READY;
-
-@ThreadSafe
-public abstract class ReasonerProducer<ANSWER> implements Producer<ANSWER>, ReasonerConsumer<ANSWER> {
-
-    private static final Logger LOG = LoggerFactory.getLogger(ReasonerProducer.class);
-
-    private final ControllerRegistry controllerRegistry;
-    final AtomicInteger requiredAnswers;
-    final Options.Query options;
-    final ExplainablesManager explainablesManager;
-    private Actor.Driver<? extends AbstractProcessor<?, ANSWER, ?, ?>> rootProcessor;
-    private Throwable exception;
-    Queue<ANSWER> queue;
-    State state;
-
-    enum State {
-        INIT,
-        INITIALISING,
-        READY,
-        PULLING,
-        FINISHED,
-        EXCEPTION
-    }
-
-    // TODO: this class should not be a Producer, it implements a different async processing mechanism
-    private ReasonerProducer(Options.Query options, ControllerRegistry controllerRegistry,
-                             ExplainablesManager explainablesManager) {
-        this.options = options;
-        this.controllerRegistry = controllerRegistry;
-        this.explainablesManager = explainablesManager;
-        this.queue = null;
-        this.requiredAnswers = new AtomicInteger();
-        this.state = INIT;
-    }
-
-    ControllerRegistry controllerRegistry() {
-        return controllerRegistry;
-    }
-
-    @Override
-    public synchronized void produce(Queue<ANSWER> queue, int requestedAnswers, Executor executor) {
-        assert this.queue == null || this.queue == queue;
-        assert requestedAnswers > 0;
-        if (state == EXCEPTION) queue.done(exception);
-        else if (state == FINISHED) queue.done();
-        else {
-            this.queue = queue;
-            requiredAnswers.addAndGet(requestedAnswers);
-            if (state == INIT) initialise();
-            else if (state == READY) pull();
-        }
-    }
-
-    private void initialise() {
-        assert state == INIT;
-        state = INITIALISING;
-        initialiseRootController();
-    }
-
-    abstract void initialiseRootController();
-
-    @Override
-    public synchronized void setRootProcessor(Actor.Driver<? extends AbstractProcessor<?, ANSWER, ?, ?>> rootProcessor) {
-        assert this.rootProcessor == null;
-        this.rootProcessor = rootProcessor;
-        state = READY;
-        if (requiredAnswers.get() > 0) pull();
-    }
-
-    synchronized void pull() {
-        assert state == READY;
-        state = PULLING;
-        rootProcessor.execute(actor -> actor.rootPull());
-    }
-
-    @Override
-    public synchronized void finish() {
-        // note: root resolver calls this single-threaded, so is thread safe
-        LOG.trace("All answers found.");
-        if (state != FINISHED && state != EXCEPTION) {
-            if (queue == null) {
-                assert state != PULLING;
-                assert requiredAnswers.get() == 0;
-            } else {
-                requiredAnswers.set(0);
-                queue.done();
-            }
-        }
-    }
-
-    @Override
-    public synchronized void exception(Throwable e) {
-        if (state != FINISHED && state != EXCEPTION) {
-            exception = e;
-            if (queue == null) {
-                assert state != PULLING;
-                assert requiredAnswers.get() == 0;
-            } else {
-                requiredAnswers.set(0);
-                queue.done(e.getCause());
-            }
-        }
-    }
-
-    @Override
-    public void recycle() {
-
-    }
-
-    public static abstract class Match extends ReasonerProducer<ConceptMap> {
-
-        Match(Options.Query options, ControllerRegistry controllerRegistry, ExplainablesManager explainablesManager) {
-            super(options, controllerRegistry, explainablesManager);
-        }
-
-        @Override
-        public synchronized void receiveAnswer(ConceptMap answer) {
-            state = READY;
-            // TODO: The explainables can always be given. The only blocked to removing the explain flag is that
-            //  `match get` filters should be ignored for an explainable answer
-            if (options.explain() && !answer.explainables().isEmpty()) {
-                explainablesManager.setAndRecordExplainables(answer);
-            }
-            queue.put(answer);
-            if (requiredAnswers.decrementAndGet() > 0) pull();
-        }
-
-        public static class Conjunction extends Match {
-
-            private final com.vaticle.typedb.core.pattern.Conjunction conjunction;
-            private final Modifiers.Filter filter;
-
-            public Conjunction(com.vaticle.typedb.core.pattern.Conjunction conjunction,
-                               Modifiers.Filter filter, Options.Query options,
-                               ControllerRegistry controllerRegistry, ExplainablesManager explainablesManager) {
-                super(options, controllerRegistry, explainablesManager);
-                this.conjunction = conjunction;
-                this.filter = filter;
-            }
-
-            @Override
-            protected synchronized void initialiseRootController() {
-                controllerRegistry().createRootConjunction(ResolvableConjunction.of(conjunction), filter, options.explain(), this);
-            }
-        }
-
-        public static class Disjunction extends Match {
-
-            private final com.vaticle.typedb.core.pattern.Disjunction disjunction;
-            private final Modifiers.Filter filter;
-
-            public Disjunction(com.vaticle.typedb.core.pattern.Disjunction disjunction,
-                               Modifiers.Filter filter, Options.Query options,
-                               ControllerRegistry controllerRegistry, ExplainablesManager explainablesManager) {
-                super(options, controllerRegistry, explainablesManager);
-                this.disjunction = disjunction;
-                this.filter = filter;
-            }
-
-            @Override
-            protected synchronized void initialiseRootController() {
-                controllerRegistry().createRootDisjunction(ResolvableDisjunction.of(disjunction), filter, options.explain(), this);
-            }
-        }
-    }
-
-    public static class Explain extends ReasonerProducer<Explanation> {
-
-        private final Concludable concludable;
-        private final ConceptMap bounds;
-
-        Explain(Concludable concludable, ConceptMap bounds, Options.Query options,
-                ControllerRegistry controllerRegistry,
-                ExplainablesManager explainablesManager) {
-            super(options, controllerRegistry, explainablesManager);
-            this.concludable = concludable;
-            this.bounds = bounds;
-        }
-
-        @Override
-        synchronized void initialiseRootController() {
-            controllerRegistry().createExplainableRoot(concludable, bounds, this);
-        }
-
-        @Override
-        public synchronized void receiveAnswer(Explanation explanation) {
-            state = READY;
-            if (!explanation.conditionAnswer().explainables().isEmpty()) {
-                explainablesManager.setAndRecordExplainables(explanation.conditionAnswer());
-            }
-            queue.put(explanation);
-            if (requiredAnswers.decrementAndGet() > 0) pull();
-        }
-    }
-}
